% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlserver_connection.R
\name{create_sqlserver_connection}
\alias{create_sqlserver_connection}
\title{Create a connection to a SQL Server database.}
\usage{
create_sqlserver_connection(server, database, timeout = 10)
}
\arguments{
\item{server}{Server and instance where SQL Server database found.}

\item{database}{The name of the database to connect to.}

\item{timeout}{The timeout period (in seconds) for establishing
the connection. Defaults to 10.}
}
\value{
A connection object of class \code{"Microsoft SQL Server"}
from the \code{odbc} package.
}
\description{
Establishes a connection to a SQL Server database
using the ODBC driver. It uses Windows authentication. Use
\code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} to disconnect this connection once no longer in use.
}
\examples{
\dontrun{
# Connect to a SQL Server database
con <- create_sqlserver_connection(
  server = "my_server",
  database = "my_database"
)

# Remember to disconnect after usage
DBI::dbDisconnect(con)
}

}
