% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_column.R
\name{drop_column}
\alias{drop_column}
\title{Drop a Column from an existing database table}
\usage{
drop_column(server, database, schema, table_name, column_name)
}
\arguments{
\item{server}{Server and instance where SQL Server database found.}

\item{database}{Database containing the table with column to drop.}

\item{schema}{Name of schema containing the table.}

\item{table_name}{Name of the table from which the column should be dropped.}

\item{column_name}{The name of the column to be dropped.}
}
\description{
Drops a specified column from a table.
Checks if the table exists in the schema and if the column is present in the
table before attempting to drop the column.
}
\examples{
\dontrun{
# Drop the Species column from test_iris table
drop_column(
  server = "my_server",
  database = "my_database",
  schema = "my_schema",
  table_name = "test_iris",
  column_name = "Species"
)
}
}
