% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_table.R
\name{drop_table_from_db}
\alias{drop_table_from_db}
\title{Drop SQL Server table from database. Check if versioned table and disable versioning and drop history table too if so.}
\usage{
drop_table_from_db(database, server, schema, table_name)
}
\arguments{
\item{database}{Database containing the table to be dropped.}

\item{server}{Server and instance where SQL Server database found.}

\item{schema}{Name of schema containing table to be dropped.}

\item{table_name}{Name of the table to be dropped.}
}
\description{
Drop SQL Server table from database. Check if versioned table and disable versioning and drop history table too if so.
}
\examples{
drop__table_from_db(database = "my_database", server = "my_server", schema = "my_schema", table_name = "table_to_drop")
}
