% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_table.R
\name{drop_versioned_table_from_db}
\alias{drop_versioned_table_from_db}
\title{Drop versioned SQL Server table from database}
\usage{
drop_versioned_table_from_db(database, server, schema, table_name)
}
\arguments{
\item{database}{Database containing the table to be dropped.}

\item{server}{Server and instance where SQL Server database found.}

\item{schema}{Name of schema containing table to be dropped.}

\item{table_name}{Name of the table to be dropped.}
}
\description{
Drop versioned SQL Server table from database
}
\examples{
drop_versioned_table_from_db(database = "my_database", server = "my_server", schema = "my_schema", table_name = "table_to_drop")
}
