% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_sql.R
\name{execute_sql}
\alias{execute_sql}
\title{Connect, execute SQL in SQL Server database and then disconnect}
\usage{
execute_sql(server, database, sql, output = FALSE)
}
\arguments{
\item{server}{Server instance where SQL Server database running.}

\item{database}{SQL Server database in which SQL executed.}

\item{sql}{SQL to be executed in the database.}

\item{output}{If TRUE write output of SQL to Dataframe. Defaults to FALSE.}
}
\value{
Dataframe or NULL depending on SQL executed.
}
\description{
Connect, execute SQL in SQL Server database and then disconnect
}
\examples{
\dontrun{
sql_to_run <- "select test_column, other_column from my_test_table
where other_column > 10"
execute_sql(
  database = "my_database", server = "my_server",
  sql = sql_to_run, output = TRUE
)
}
}
