% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{rename_column}
\alias{rename_column}
\title{Rename a column in an existing database table.}
\usage{
rename_column(
  server,
  database,
  schema,
  table_name,
  old_column_name,
  new_column_name
)
}
\arguments{
\item{server}{Server and instance where SQL Server database found.}

\item{database}{Database containing the table in which the column
belongs.}

\item{schema}{Name of the schema containing the table.}

\item{table_name}{Name of the table containing the column.}

\item{old_column_name}{The current name of the column to be renamed.}

\item{new_column_name}{The new name for the column.}
}
\description{
Renames a specified column in a table. The function checks if the table
and the column exist in the schema before attempting to rename the column.
In MS SQL Server this is done using \verb{EXEC sp_rename}.
}
\examples{
\dontrun{
# Rename Species column to SpeciesRenamed in test_iris table
rename_column(
  server = "my_server",
  database = "my_database",
  schema = "my_schema",
  table_name = "test_iris",
  old_column_name = "Species",
  new_column_name = "SpeciesRenamed"
)
}
}
