% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_table.R
\name{rename_table}
\alias{rename_table}
\title{Rename a table in the database}
\usage{
rename_table(server, database, schema, old_table_name, new_table_name)
}
\arguments{
\item{server}{Server and instance where SQL Server database found.}

\item{database}{Database containing the table to be renamed.}

\item{schema}{Name of the schema containing the table.}

\item{old_table_name}{The current name of the table to be renamed.}

\item{new_table_name}{The new name for the table.}
}
\description{
Renames a specified table. The function checks if the table exists
in the schema before attempting to rename it.
In MS SQL Server, this is done using \verb{EXEC sp_rename}.
}
\examples{
\dontrun{
# Rename test_iris table to test_iris_renamed
rename_table(
  server = "my_server",
  database = "my_database",
  schema = "my_schema",
  old_table_name = "test_iris",
  new_table_name = "test_iris_renamed"
)
}
}
