% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_replace_table.R
\name{write_dataframe_to_db}
\alias{write_dataframe_to_db}
\title{Write an R dataframe to SQL Server table with system versioning.}
\usage{
write_dataframe_to_db(database, server, schema, table_name, dataframe)
}
\arguments{
\item{database}{Name of SQL Server database where table will be written.}

\item{server}{Server instance where SQL Server database running.}

\item{schema}{Name of schema in SQL Server database where table will be created.}

\item{table_name}{Name of table to be created in SQL Server database.}

\item{dataframe}{Source R dataframe that will be written to SQL Server database.}
}
\description{
Write an R dataframe to SQL Server table with system versioning.
}
\examples{
write_dataframe_to_db(database = "my_database", server = "my_server", schema = "my_schema", table_name = "output_table", dataframe = my_df)
}
