% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column.R
\name{add_column}
\alias{add_column}
\title{Add a Column to an existing database table}
\usage{
add_column(
  server,
  database,
  schema,
  table_name,
  column_name,
  sql_data_type = NULL,
  sample_value = NULL
)
}
\arguments{
\item{server}{Server and instance where SQL Server database found.}

\item{database}{Database containing the table to which the column
will be added.}

\item{schema}{Name of schema containing the table.}

\item{table_name}{Name of the table to which the column should be added.}

\item{column_name}{The name of the column to be added.}

\item{sql_data_type}{The SQL datatype for the column as a character string.
For example \code{"nvarchar(255)"}
(optional, can be inferred from \code{sample_value} alternatively). Use
\code{\link[=db_table_metadata]{db_table_metadata()}} to see the data types for existing tables/columns
in the database, or refer to MS SQL Server guidance on data types.}

\item{sample_value}{Existing R data frame column,
or a value that defines the datatype of the column. The input
should be of the correct R class for its type, for example, strings as
character, numbers as \code{numeric}, dates as \code{Date}, date times as \code{POSIXct}
or \code{POSIXlt}.
Optional and superseded by \code{sql_data_type} if specified. Some example
valid inputs: \code{iris$Species}, \code{"a sample string"}, \code{numeric()},
\verb{as.Date("2024-01-01"}).}
}
\description{
Adds a specified column to a table. The column data type can be mapped
from an R object, or else the database column data type must be
explicitly specified.
}
\examples{
\dontrun{
# Add a Species_new column to test_iris table
add_column(
  server = "my_server",
  database = "my_database",
  schema = "my_schema",
  table_name = "test_iris",
  column_name = "Species_new",
  sample_value = character(0)
)
}
}
