% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_sql}
\alias{execute_sql}
\title{Create a connection to SQL Server database, execute SQL and then (optionally) disconnect from database.}
\usage{
execute_sql(database, server, sql, output = FALSE, disconnect = TRUE)
}
\arguments{
\item{database}{SQL Server database in which SQL executed.}

\item{server}{Server instance where SQL Server database running.}

\item{sql}{SQL to be executed in the database.}

\item{output}{If TRUE write output of SQL to Dataframe. Defaults to FALSE.}

\item{disconnect}{If TRUE disconnect from SQL Server database at end. Defaults to TRUE.}
}
\value{
Dataframe or NULL depending on SQL executed.
}
\description{
Create a connection to SQL Server database, execute SQL and then (optionally) disconnect from database.
}
\examples{
sql_to_run <- "select test_column, other_column from my_test_table where other_column > 10"
execute_sql(database=my_database, server=my_server, sql=sql_to_run, output=TRUE, disconnect=TRUE)
}
