% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table_from_db}
\alias{read_table_from_db}
\title{Read a SQL Server table into an R dataframe.}
\usage{
read_table_from_db(
  database,
  server,
  schema,
  table_name,
  columns = NULL,
  filter_stmt = NULL
)
}
\arguments{
\item{database}{Database containing table to read.}

\item{server}{Server instance where SQL Server database running.}

\item{schema}{Name of database schema containing table to read.}

\item{table_name}{Name of table in database to read.}

\item{columns}{Optional vector of column names to select.}

\item{filter_stmt}{Optional filter statement - this should be a character
expression in the format of a \code{\link[dplyr:filter]{dplyr::filter()}} query,
for example \code{"Species == 'virginica'"} and it will be translated to SQL
using \code{\link[dbplyr:translate_sql]{dbplyr::translate_sql()}}.}
}
\value{
Dataframe of table.
}
\description{
If you are confident in writing SQL you may prefer to
use the \code{\link[=execute_sql]{execute_sql()}} function instead.
}
\examples{
\dontrun{
read_table_from_db(
  database = "my_database",
  server = "my_server",
  table_name = "my_table",
  columns = c("column1", "column2"),
  filter_stmt = "column1 < 5 & column2 == 'b'")
)
}
}
