% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table_from_db}
\alias{read_table_from_db}
\title{Read a SQL Server table to R dataframe,}
\usage{
read_table_from_db(database, server, schema, table_name, columns = NULL)
}
\arguments{
\item{database}{Database containing table to read.}

\item{server}{Server instance where SQL Server database running.}

\item{schema}{Name of database schema containing table to read.}

\item{table_name}{Name of table in database to read.}

\item{columns}{Optional vector of column names to select.}
}
\value{
Dataframe of table.
}
\description{
Read a SQL Server table to R dataframe,
}
\examples{
read_table_from_db(database = "my_database", server = "my_server", table_name = "my_table", columns = c("column1", "column2"))
}
